Update: - Change Frequency from long to double - Implement Data Annotation on GUI - Change some default value complied with new discussion - Change binding UI - minor change on yml, readme (thank to Dennis)

+++ WpfApp2/WpfApp2/ViewModel/GenerateSignalData.cs
@@ -32,7 +32,7 @@ namespace WpfApp2.ViewModel
         [JsonIgnore]
         public WaveForm mWave { get; set; }
         [JsonIgnore]
-        public long mFreq { get; set; }
+        public double mFreq { get; set; }
         [JsonIgnore]
         public  double mAmpl { get; set; }
         [JsonIgnore]
@@ -133,7 +133,7 @@ namespace WpfApp2.ViewModel
             return mProfile;
         }
 
-        public long getFreq()
+        public double getFreq()
         {
             return mFreq;
         }
